fix: ensure mobile menu & search toggle respond
Fixes a problem where the mobile menu did not respond visually (did not
look open) if the search was open, due to the fact that the search has a
higher z-index. When fixing that, then the menu would not show when
clicked if it was already open and the search was open too, as it would
be closing.

This fixes both issues and makes it so whenever the search or menu
toggle are clicked, it closes the other overlay and shows only the
corresponding one.